Experiment /w a caching commit view
Trying to fix #1, the high level idea here is to only walk the revision
graph once, as the graph is walked then commits are also stored in a
cache so that if the user goes back up the history we don't have to
start a new walker. Instead older commits are served from the cache.

This would mean we could probably remove the startup tree scan for
matching commits entirely at the cost of growing memory usage over time.

However, I painted myself into a corner /w the borrow checker and
probably need to revisit an abstraction here.