Skip commits /w the same tree when using path filtering
When tree's are the same, they can be immediately skipped. This should
improve start-up times (faster to calculate all commits where the trees
changed, than when the diff contains the path), at the expense of longer
load times when walking the commits (have to check the diffs that were
not filtered).